AppDelegate, SceneDelegate整理(テストMock化)
